pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rnaseq_mini_pipeline"
version = "0.1.0"
description = "A mini RNA-seq data processing pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    { name = "Laroussi Bachri", email = "bachri.laroussi-labid@utoulouse.fr" }
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "seaborn>=0.11.0",
    "matplotlib>=3.4.0",
    "scikit-learn>=1.2.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0",
    "PyYAML>=5.4.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project. optional-dependencies]
dev = [
    "pytest>=6.2.0"
]

[project.scripts]
rnaseq-pipeline = "rnaseq.pipeline:main"

[project.urls]
Homepage = "https://github.com/bachrilrs/rnaseq_mini_pipeline"
Repository = "https://github.com/bachrilrs/rnaseq_mini_pipeline"
LinkedIn = "https://www.linkedin.com/in/laroussi-bachri"